+++ elastic_db_repo.py
@@ -80,7 +80,6 @@ import os
 
 # Local
 try:
-    from .lib import arg_parser
     from .lib import gen_libs
     from .lib import gen_class
     from .elastic_lib import elastic_class
@@ -88,7 +87,6 @@ try:
     from . import version
 
 except (ValueError, ImportError) as err:
-    import lib.arg_parser as arg_parser
     import lib.gen_libs as gen_libs
     import lib.gen_class as gen_class
     import elastic_lib.elastic_class as elastic_class
@@ -124,9 +122,9 @@ def list_dumps(els, **kwargs):
         repositories.
 
     Arguments:
-        (input) els -> ElasticSearch class instance.
+        (input) els -> ElasticSearch class instance
         (input) **kwargs:
-            args_array -> Dict of command line options and values.
+            args -> ArgParser class instance
 
     """
 
@@ -155,23 +153,23 @@ def create_repo(els, repo_name=None, repo_dir=None, **kwargs):
     Description:  Create an ElasticSearch repository.
 
     Arguments:
-        (input) els -> ElasticSearch class instance.
-        (input) repo_name -> Name of repository.
-        (input) repo_dir -> Repository directory path.
+        (input) els -> ElasticSearch class instance
+        (input) repo_name -> Name of repository
+        (input) repo_dir -> Repository directory path
         (input) **kwargs:
-            args_array -> Dict of command line options and values.
+            args -> ArgParser class instance
 
     """
 
     global PRT_TEMPLATE
 
-    args_array = dict(kwargs.get("args_array"))
+    args = kwargs.get("args")
 
     if not repo_name:
-        repo_name = args_array.get("-C")
+        repo_name = args.get_val("-C")
 
     if not repo_dir:
-        repo_dir = args_array.get("-l")
+        repo_dir = args.get_val("-l")
 
     if repo_name in els.repo_dict:
         print("Error:  '%s' repository already exists at: '%s'"
@@ -194,20 +192,20 @@ def delete_repo(els, repo_name=None, **kwargs):
     Description:  Delete an Elasticsearch repository.
 
     Arguments:
-        (input) els -> ElasticSearch class instance.
-        (input) repo_name -> Name of repository.
+        (input) els -> ElasticSearch class instance
+        (input) repo_name -> Name of repository
         (input) **kwargs:
-            args_array -> Dict of command line options and values.
+            args -> ArgParser class instance
 
     """
 
     global WARN_TEMPLATE
     global PRT_TEMPLATE
 
-    args_array = dict(kwargs.get("args_array"))
+    args = kwargs.get("args")
 
     if not repo_name:
-        repo_name = args_array.get("-D")
+        repo_name = args.get_val("-D")
 
     if repo_name in els.repo_dict:
 
@@ -228,24 +226,24 @@ def delete_dump(els, repo_name=None, dump_name=None, **kwargs):
     Description:  Delete a dump in an Elasticsearch repository.
 
     Arguments:
-        (input) els -> ElasticSearch class instance.
-        (input) repo_name -> Name of repository.
-        (input) dump_name -> Name of dump to delete.
+        (input) els -> ElasticSearch class instance
+        (input) repo_name -> Name of repository
+        (input) dump_name -> Name of dump to delete
         (input) **kwargs:
-            args_array -> Dict of command line options and values.
+            args -> ArgParser class instance
 
     """
 
     global WARN_TEMPLATE
     global PRT_TEMPLATE
 
-    args_array = dict(kwargs.get("args_array"))
+    args = kwargs.get("args")
 
     if not repo_name:
-        repo_name = args_array.get("-r")
+        repo_name = args.get_val("-r")
 
     if not dump_name:
-        dump_name = args_array.get("-S")
+        dump_name = args.get_val("-S")
 
     if repo_name in els.repo_dict:
 
@@ -276,17 +274,17 @@ def rename_repo(els, name_list=None, **kwargs):
     Description:  Rename an Elasticseatch repository.
 
     Arguments:
-        (input) els -> ElasticSearch class instance.
-        (input) name_list -> List of two repository names for renaming process.
+        (input) els -> ElasticSearch class instance
+        (input) name_list -> List of two repository names for renaming process
         (input) **kwargs:
-            args_array -> Dict of command line options and values.
+            args -> ArgParser class instance
 
     """
 
-    args_array = dict(kwargs.get("args_array"))
+    args = kwargs.get("args")
 
     if not name_list:
-        name_list = list(args_array.get("-M"))
+        name_list = list(args.get_val("-M"))
 
     if isinstance(name_list, list) and len(name_list) == 2:
 
@@ -315,8 +313,8 @@ def _rename(els, name_list):
     Description:  Private function for rename_repo function.
 
     Arguments:
-        (input) els -> ElasticSearch class instance.
-        (input) name_list -> List of two repository names for renaming process.
+        (input) els -> ElasticSearch class instance
+        (input) name_list -> List of two repository names for renaming process
 
     """
 
@@ -347,9 +345,9 @@ def disk_usage(els, **kwargs):
     Description:  Display disk usage of dump partitions.
 
     Arguments:
-        (input) els -> ElasticSearch class instance.
+        (input) els -> ElasticSearch class instance
         (input) **kwargs:
-            args_array -> Dict of command line options and values.
+            args -> ArgParser class instance
 
     """
 
@@ -377,16 +375,16 @@ def list_repos(els, **kwargs):
     Description:  Lists the repositories present.
 
     Arguments:
-        (input) els -> ElasticSearch class instance.
+        (input) els -> ElasticSearch class instance
         (input) **kwargs:
-            args_array -> Dict of command line options and values.
+            args -> ArgParser class instance
 
     """
 
     elastic_libs.list_repos2(els.repo_dict)
 
 
-def run_program(args_array, func_dict):
+def run_program(args, func_dict):
 
     """Function:  run_program
 
@@ -394,15 +392,13 @@ def run_program(args_array, func_dict):
         Create a program lock to prevent other instantiations from running.
 
     Arguments:
-        (input) args_array -> Dict of command line options and values.
-        (input) func_dict -> Dictionary list of functions and options.
+        (input) args -> ArgParser class instance
+        (input) func_dict -> Dictionary list of functions and options
 
     """
 
-    cmdline = gen_libs.get_inst(sys)
-    args_array = dict(args_array)
     func_dict = dict(func_dict)
-    cfg = gen_libs.load_module(args_array["-c"], args_array["-d"])
+    cfg = gen_libs.load_module(args.get_val("-c"), args.get_val("-d"))
     user = cfg.user if hasattr(cfg, "user") else None
     japd = cfg.japd if hasattr(cfg, "japd") else None
     ca_cert = cfg.ssl_client_ca if hasattr(cfg, "ssl_client_ca") else None
@@ -410,17 +406,17 @@ def run_program(args_array, func_dict):
     flavorid = "elasticrepo"
 
     try:
-        prog_lock = gen_class.ProgramLock(cmdline.argv, flavor_id=flavorid)
+        prog_lock = gen_class.ProgramLock(sys.argv, flavor_id=flavorid)
 
         # Find which functions to call.
-        for opt in set(args_array.keys()) & set(func_dict.keys()):
+        for opt in set(args.get_args_keys()) & set(func_dict.keys()):
             els = elastic_class.ElasticSearchRepo(
-                cfg.host, port=cfg.port, repo=args_array.get("-L"),
+                cfg.host, port=cfg.port, repo=args.get_val("-L"),
                 user=user, japd=japd, ca_cert=ca_cert, scheme=scheme)
             els.connect()
 
             if els.is_connected:
-                func_dict[opt](els, args_array=args_array)
+                func_dict[opt](els, args=args)
 
             else:
                 print("ERROR:  Failed to connect to Elasticsearch")
@@ -439,45 +435,48 @@ def main():
         line arguments and values.
 
     Variables:
-        dir_chk_list -> contains options which will be directories.
-        func_dict -> dictionary list for the function calls or other options.
-        opt_con_req_dict -> contains options requiring other options.
-        opt_multi_list -> contains the options that will have multiple values.
-        opt_req_list -> contains options that are required for the program.
-        opt_val -> List of options that allow 0 or 1 value for option.
-        opt_val_list -> contains options which require values.
-        opt_xor_dict -> contains dict with key that is xor with it's values.
+        dir_perms_chk -> contains options which will be directories and the
+            octal permission settings
+        func_dict -> dictionary list for the function calls or other options
+        opt_con_req_dict -> contains options requiring other options
+        opt_multi_list -> contains the options that will have multiple values
+        opt_req_list -> contains options that are required for the program
+        opt_val_bin -> List of options that allow 0 or 1 value for option
+        opt_val -> contains options which require values
+        opt_xor_dict -> contains dict with key that is xor with it's values
 
     Arguments:
-        (input) argv -> Arguments from the command line.
+        (input) argv -> Arguments from the command line
 
     """
 
-    cmdline = gen_libs.get_inst(sys)
-    dir_chk_list = ["-d"]
-    func_dict = {"-L": list_dumps, "-R": list_repos, "-C": create_repo,
-                 "-D": delete_repo, "-S": delete_dump, "-M": rename_repo,
-                 "-U": disk_usage}
+    dir_perms_chk = {"-d": 5}
+    func_dict = {
+        "-L": list_dumps, "-R": list_repos, "-C": create_repo,
+        "-D": delete_repo, "-S": delete_dump, "-M": rename_repo,
+        "-U": disk_usage}
     opt_con_req_dict = {"-C": ["-l"], "-S": ["-r"]}
     opt_multi_list = ["-M"]
     opt_req_list = ["-c", "-d"]
-    opt_val = ["-L"]
-    opt_val_list = ["-c", "-d", "-C", "-l", "-D", "-S", "-r", "-M"]
-    opt_xor_dict = {"-C": ["-L", "-R", "-S", "-D", "-M", "-U"],
-                    "-D": ["-L", "-R", "-S", "-C", "-M", "-U"],
-                    "-S": ["-L", "-R", "-C", "-D", "-M", "-U"],
-                    "-M": ["-L", "-R", "-S", "-C", "-D", "-U"]}
+    opt_val_bin = ["-L"]
+    opt_val = ["-c", "-d", "-C", "-l", "-D", "-S", "-r", "-M"]
+    opt_xor_dict = {
+        "-C": ["-L", "-R", "-S", "-D", "-M", "-U"],
+        "-D": ["-L", "-R", "-S", "-C", "-M", "-U"],
+        "-S": ["-L", "-R", "-C", "-D", "-M", "-U"],
+        "-M": ["-L", "-R", "-S", "-C", "-D", "-U"]}
 
     # Process argument list from command line.
-    args_array = arg_parser.arg_parse2(
-        cmdline.argv, opt_val_list, opt_val=opt_val, multi_val=opt_multi_list)
+    args = gen_class.ArgParser(
+        sys.argv, opt_val=opt_val, opt_val_bin=opt_val_bin,
+        multi_val=opt_multi_list, do_parse=True)
 
-    if not gen_libs.help_func(args_array, __version__, help_message) \
-       and not arg_parser.arg_require(args_array, opt_req_list) \
-       and arg_parser.arg_xor_dict(args_array, opt_xor_dict) \
-       and arg_parser.arg_cond_req_or(args_array, opt_con_req_dict) \
-       and not arg_parser.arg_dir_chk_crt(args_array, dir_chk_list):
-        run_program(args_array, func_dict)
+    if not gen_libs.help_func(args, __version__, help_message)  \
+       and args.arg_require(opt_req=opt_req_list)               \
+       and args.arg_xor_dict(opt_xor_val=opt_xor_dict)          \
+       and args.arg_cond_req_or(opt_con_or=opt_con_req_dict)    \
+       and args.arg_dir_chk(dir_perms_chk=dir_perms_chk):
+        run_program(args, func_dict)
 
 
 if __name__ == "__main__":